long is actually 64bit on amd64 Linux and MacOS!!! Gotta replace them with int.

+++ src/CompatibleSystem.cpp
@@ -10,7 +10,7 @@
 
 namespace Compatible
 {
-   unsigned long GetMilliseconds()
+   unsigned int GetMilliseconds()
    {
 #ifdef WIN32
       return timeGetTime();
