+++ src/State_Title.cpp
@@ -379,7 +379,7 @@ void TitleState::PlayDevicePreview(microseconds_t delta_microseconds)
       if (m_state.midi_in && m_state.midi_in->GetDeviceDescription().id == UINT32_MAX - 1 && (ev->Type() == MidiEventType_NoteOn || ev->Type() == MidiEventType_NoteOff)) {
          // Write midi input buffer for real!
 #ifdef WIN32
-         m_state.midi_in->InputCallback(MIM_DATA, (unsigned long(ev->StatusCode())) | (unsigned long(ev->NoteNumber()) << 8) | (unsigned long(ev->NoteVelocity()) << 16), NULL);
+         m_state.midi_in->InputCallback(MIM_DATA, (unsigned int(ev->StatusCode())) | (unsigned int(ev->NoteNumber()) << 8) | (unsigned int(ev->NoteVelocity()) << 16), NULL);
 #else
          m_state.midi_in->InputCallback(ev->StatusCode(), ev->NoteNumber(), ev->NoteVelocity());
 #endif
